Unify a11y editor dialogs
