+++ src/components/NodeToolbar/AccessibilityEditor/CustomRadio.js
@@ -8,7 +8,7 @@ import RadioButtonSelected from '../../icons/ui-elements/RadioButtonSelected';
 
 type Props = {
   currentValue?: ?string,
-  shownValue: string,
+  shownValue?: ?string,
   description: ?string,
   caption: ?string,
   children?: ?React.Node,
