+++ src/components/NodeToolbar/AccessibilityEditor/StyledRadioGroup.js
@@ -68,10 +68,7 @@ const StyledRadioGroup = styled(RadioGroup)`
     }
     footer {
       margin: 0.5em 0 0 0;
-      opacity: 0.6;
-      @media (max-width: 512px), (max-height: 512px) {
-        font-size: 80%;
-      }
+      opacity: 0.8;
     }
     cursor: pointer;
     input {
