+++ src/lib/useImperialUnits.js
@@ -1,0 +1,5 @@
+import { currentLocales } from './i18n';
+
+export default function useImperialUnits() {
+  return currentLocales[0] === 'en' || Boolean(currentLocales[0].match(/(UK|US)$/));
+}
